-añadido

+++ src/model/followingmodels/KraussModel.java
@@ -6,14 +6,14 @@ package model.followingmodels;
  */
 public class KraussModel implements FollowingModel {
 
-    private final static String FORMAT = "carFollowModel=\"%s\" sigma=\"%i\""
-            + " minGap=\"%i\" ";
+    private final static String FORMAT = "carFollowModel=\"%s\" sigma=\"%.1f\""
+            + " minGap=\"%d\" ";
     private final static String NAME = "Krauss";
     
-    private final int sigma;
+    private final double sigma;
     private final int minGap;
 
-    public KraussModel(int sigma, int minGap) {
+    public KraussModel(double sigma, int minGap) {
         if (sigma > 1) sigma = 1;
         if (sigma < 0) sigma = 0;
         this.sigma = sigma;
